requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
chardet==4.0.0
click==8.0.3
Django==3.2.8
djangorestframework==3.12.4
h11==0.12.0
idna==3.3
multidict==5.2.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
yarl==1.7.0
